backend/pyproject.toml
[project]
name = "savemoney-backend"
version = "0.1.0"
description = "SaveMoney 记账应用后端"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "langgraph>=0.0.40",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "openai>=1.3.0",
    "httpx>=0.25.2",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "lark-oapi>=1.4.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
# 使用uv进行包管理